+++ examples/resources/sdwan_zone_list_policy_object/resource.tf
@@ -1,0 +1,8 @@
+resource "sdwan_zone_list_policy_object" "example" {
+  name = "Example"
+  entries = [
+    {
+      vpn = "1"
+    }
+  ]
+}