+++ docs/guides/changelog.md
@@ -14,6 +14,7 @@ description: |-
 - Add `sdwan_ips_signature_list_policy_object` resource and data source
 - Add `sdwan_allow_url_list_policy_object` resource and data source
 - Add `sdwan_block_url_list_policy_object` resource and data source
+- Add `sdwan_zone_list_policy_object` resource and data source
 
 ## 0.2.4